# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CategoryResponse:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
...
  dtos.CreateCategoryRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        type: string
    required:
    - name
    - type
    type: object
...
  dtos.CurrencyResponseDto:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      symbol:
        type: string
      updated_at:
        type: string
    type: object
...
  dtos.ErrorResponse:
    properties:
      details:
        example: additional error details
        type: string
      error:
        example: error message
        type: string
    type: object
...
  dtos.SuccessResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
...
  dtos.UpdateCategoryRequest:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@arabella.app
    name: Arabella Support
  description: API REST para el sistema de gestión financiera personal Arabella. Implementa
    contabilidad de doble partida, cálculo de Runway y soporte multi-moneda.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Arabella Financial OS API
  version: 1.0.0
paths:
...
  /categories:
    get:
      description: Obtiene todas las categorías del usuario autenticado. Se puede
        filtrar por tipo (INCOME o EXPENSE)
      parameters:
      - description: Tipo de categoría (INCOME, EXPENSE)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de categorías
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  $ref: '#/definitions/dtos.CategoryResponse'
                type: array
            type: object
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar categorías
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: 'Crea una nueva categoría para clasificar transacciones del usuario
        autenticado. Tipos válidos: INCOME (ingresos) o EXPENSE (gastos)'
      parameters:
      - description: Datos de la nueva categoría
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Categoría creada exitosamente
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.CategoryResponse'
              message:
                type: string
            type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear categoría
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Realiza un borrado lógico (soft delete) de una categoría. Las transacciones
        asociadas no se eliminan
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categoría eliminada exitosamente
          schema:
            $ref: '#/definitions/dtos.SuccessResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar categoría
      tags:
      - Categories
    get:
      description: Obtiene el detalle de una categoría específica por su ID
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalle de la categoría
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.CategoryResponse'
            type: object
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Categoría no encontrada
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener categoría por ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Actualiza el nombre o el estado activo de una categoría existente.
        Solo se modifican los campos enviados
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Categoría actualizada exitosamente
          schema:
            $ref: '#/definitions/dtos.SuccessResponse'
        "400":
          description: ID o datos inválidos
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: No autenticado
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar categoría
      tags:
      - Categories
  /currencies:
    get:
      description: Obtiene todas las monedas activas disponibles en el sistema (USD,
        EUR, MXN, etc.). Este endpoint es público y no requiere autenticación
      produces:
      - application/json
      responses:
        "200":
          description: Lista de monedas activas
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  $ref: '#/definitions/dtos.CurrencyResponseDto'
                type: array
            type: object
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Listar monedas disponibles
      tags:
      - Currencies
  /currencies/{code}:
    get:
      description: 'Obtiene el detalle de una moneda específica usando su código ISO
        4217 (ej: USD, EUR, MXN). Este endpoint es público y no requiere autenticación'
      parameters:
      - description: 'Código ISO 4217 de la moneda (ej: USD, EUR, MXN)'
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detalle de la moneda
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.CurrencyResponseDto'
            type: object
        "404":
          description: Moneda no encontrada
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Obtener moneda por código ISO
      tags:
      - Currencies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Escribe "Bearer" seguido de un espacio y tu JWT de acceso. Ejemplo:
      "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"